support nil-metrics (#380)
* support nil-metrics

With the change to passing around an interface intead of a concrete type, it was necessary to check `metrics != nil`. Improve this by always setting `metrics` to a concrete value, even if the underlying value is `nil`. This allows all calling code to safely use `metrics` without a `nil` check.

* also handle p.metricsService != nil

* prefer p.GetMetrics

* skip != nil in middleware, since redundant

* just use metrics all the time