Metrics & naming conventions (#378)
* support nil-metrics

With the change to passing around an interface intead of a concrete type, it was necessary to check `metrics != nil`. Improve this by always setting `metrics` to a concrete value, even if the underlying value is `nil`. This allows all calling code to safely use `metrics` without a `nil` check.

* also handle p.metricsService != nil

* prefer p.GetMetrics

* skip != nil in middleware, since redundant

* drop changeEventTotal

* rename processedChangeEvent

* rename lifecycleEvent interface method

* update to match Prometheus naming conventions

* move discarded reasons to metrics/

* just use metrics all the time

* updated naming

* fix ObserveChangeEventQueueRejected naming

* make generate