MM-55507: adopt ServeMetrics hook
Adopt the new [`ServeMetrics`](https://github.com/mattermost/mattermost/pull/24249) hook to expose Prometheus metrics on `:8067/plugins/com.mattermost.msteams-sync/metrics`. This /duplicates/ the existing `:9094/metrics` for now, with the latter to be removed once we finish migrating to the new endpoint in Mattermost Cloud. This is not a breaking change.

As part of this change, we unconditionally initialize the `metricsServer` so it's never `nil`. We also now only stop the metrics server `OnDeactivate`.

Fixes: https://mattermost.atlassian.net/browse/MM-55507