don't restart activity handlers on reset
While it's safe to restart the background jobs and the activity handler, stopping the workers for every configuration change (even if unrelated to the plugin!) adds churn to the primary messge processing pipeline. Since the handler itself is stateless relative to the configuration -- it always looks up the latest configuration as needed -- we can skip it during restart and only stop `OnDeactivate`.

Fixes: https://mattermost.atlassian.net/browse/MM-55490