Skip re-connect of primary app client
In the course of making the changes in this pull request, I realized that we don't actually use any new `TenantID`, `ClientID` or `ClientSecret` on configuration change: so the re`Connect()` on the primary app client is redundant. It might also be dangerous, if the client is briefly unusable during the reconnection.