MM-55490: dont reset on configuration changes (#445)
* don't restart activity handlers on reset

While it's safe to restart the background jobs and the activity handler, stopping the workers for every configuration change (even if unrelated to the plugin!) adds churn to the primary messge processing pipeline. Since the handler itself is stateless relative to the configuration -- it always looks up the latest configuration as needed -- we can skip it during restart and only stop `OnDeactivate`.

Fixes: https://mattermost.atlassian.net/browse/MM-55490

* Skip re-connect of primary app client

In the course of making the changes in this pull request, I realized that we don't actually use any new `TenantID`, `ClientID` or `ClientSecret` on configuration change: so the re`Connect()` on the primary app client is redundant. It might also be dangerous, if the client is briefly unusable during the reconnection.