Remove enforce connected flow
This feature was intended to force users to connect to MS Teams, but as this is no longer part of the rollout plan, let’s remove it to reduce surface area and reduce risk. It also has a non-trivial impact on plugin updates when all connected users smash the endpoint.

Fixes: https://mattermost.atlassian.net/browse/MM-56756