avoid assets error
When this plugin is deployed, the `assets/` folder ends up being scanned within the monorepo as part of `make check-style` and spits out the following:

```
plugins/com.mattermost.msteams-sync/assets/assets.go:1:1: License not found or wrong
make: *** [vet] Error 1
```

That's because the `assets/` folder both gets copied with the plugin **and** is used to compile assets directly into the binary. Let's simplify this and just read the files from the `assets/` folder (like we do for profile pictures), and revisit the user of embedding in a future iteration that abandons the `assets/` copying strategy altogether.