MM-56811: notify when attachments not synced
Don't surprise end users if file attachments are disabled by simply ignoring them: send a message (inside Mattermost) if a file attachment from Teams is skipped or if a user tries to send an attachment from Mattermost.

This is not sent as an ephemeral message, partly since it applies to all channel participants, but also because an ephemeral message will disappear on reload.

Fixes: https://mattermost.atlassian.net/browse/MM-56811