Remove FOR UPDATE from read-only transactions (#676)
In the past, we relied on transactions at the application layer, and had embedded `FOR UPDATE` additions to various queries in support of same. This stopped working once we added support for read replicas and marked these APIs with `db:withReplica`. But notably, this is only an issue when a read replica is actually setup and the database enforces read-only semantics.