skip assets.go when building (#814)
Don't bundle `assets/asset.go` into the final plugin distribution: this
is used exclusively to compile certain assets into the resulting binary,
but when included in `assets/` it gets copied into the server's plugin
directory. In production, this is harmless, but for developers, this
trips up `make check-style` which sees the file doesn't conform with the
monorepo style.